refac: alterado a construção do método TimerController; tomara que dessa vez de certo.
